Add MAX_PIDS subsetting, case2 graceful fallback, AMP/compile flags, QFT memory

+++ quantum_models/quantum_temporal_diff.py
@@ -145,7 +145,7 @@ class QuantumTemporalDiff(nn.Module):
         angles_cpu  = angles.permute(1, 0, 2).cpu().float()  # [T-1, B, n_q]
         weights_cpu = self.qlayer_weights.cpu().float()
 
-        q_out = self.circuit(angles_cpu, weights_cpu).to(input_device)  # [B, 2^n_q]
+        q_out = self.circuit(angles_cpu, weights_cpu).float().to(input_device)  # [B, 2^n_q]
 
         delta = self.upscale(q_out)   # [B, in_features]
